Project updated.  minSdkVersion changed to 22, targetSdkVersion 29, libraries updated to androidx. All warnings have been assessed and corrected if necessary.

+++ app/src/main/java/com/clacksdepartment/hype/SettingsFragment.java
@@ -3,35 +3,45 @@ package com.clacksdepartment.hype;
 import android.content.SharedPreferences;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
-import android.preference.Preference;
-import android.preference.PreferenceFragment;
-import android.preference.PreferenceManager;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
+import androidx.preference.PreferenceManager;
 
-public class SettingsFragment extends PreferenceFragment {
+public class SettingsFragment extends PreferenceFragmentCompat  {
+    @Override
+    public void onCreatePreferences(Bundle bundle, String s) {
+        // Load the Preferences from the XML file
+        addPreferencesFromResource(R.xml.preferences);
+    }
+
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        // Load the preferences from an XML resource
-        addPreferencesFromResource(R.xml.preferences);
         Preference button = findPreference("pref_db");
-        button.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                // Drop DB
-                SQLiteDatabase db = (new FeedReaderDbHelper(getActivity().getApplicationContext())).getWritableDatabase();
-                db.delete(FeedReaderContract.FeedEntryReleases.TABLE_NAME, null, null);
+        if (button != null)
+            button.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                @Override
+                public boolean onPreferenceClick(Preference preference) {
+                    // Drop DB
+                    SQLiteDatabase db;
+                    if (getActivity() != null){
+                        db = (new FeedReaderDbHelper(getActivity().getApplicationContext())).getWritableDatabase();
+                        db.delete(FeedReaderContract.FeedEntryReleases.TABLE_NAME, null, null);
 
-                // Change pref_db to notify mModifiedListAdapter
-                SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getActivity().getApplicationContext());
-                boolean value = prefs.getBoolean("pref_db",false);
-                SharedPreferences.Editor editor = prefs.edit();
-                editor.putBoolean("pref_db", !value);
-                editor.apply();
+                        // Change pref_db to notify mModifiedListAdapter
+                        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getActivity().getApplicationContext());
+                        boolean value = prefs.getBoolean("pref_db",false);
+                        SharedPreferences.Editor editor = prefs.edit();
+                        editor.putBoolean("pref_db", !value);
+                        editor.apply();
 
-                return true;
-            }
-        });
+                        return true;
+                    } else {
+                        return false;
+                    }
+                }
+            });
 
     }
 }
